Ajout Code promo

+++ src/Entity/Livreur.php
@@ -31,7 +31,7 @@ class Livreur
     private $description;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $tarif;
 
@@ -74,12 +74,12 @@ class Livreur
         return $this;
     }
 
-    public function getTarif(): ?int
+    public function getTarif(): ?float
     {
         return $this->tarif;
     }
 
-    public function setTarif(int $tarif): self
+    public function setTarif(float $tarif): self
     {
         $this->tarif = $tarif;
